+++ js/config.js
@@ -6,7 +6,7 @@
 
 export const APP_CONFIG = Object.freeze({
   name: 'PrincessCycle',
-  version: '1.1.0',
+  version: '2.0.0',
 
   // Storage
   dbName: 'PrincessCycleDB',